refactor(cli): move table printer into Texture
